requirements.txt
Babel==1.3
Beaker==1.6.4
FormEncode==1.3.0a1
Genshi==0.7
Mako==0.9.1
MarkupSafe==0.23
PasteDeploy==1.5.2
SQLAlchemy==0.9.4
Tempita==0.5.3dev
TurboGears2==2.3.2
WebOb==1.3.1
WebTest==1.4.3
alembic==0.6.5
argparse==1.2.1
backlash==0.0.5
cliff==1.6.1
cmd2==0.6.7
crank==0.7.0
decorator==3.4.0
docutils==0.11
gearbox==0.0.6
prettytable==0.7.2
pyparsing==2.0.2
pytz==2014.2
repoze.lru==0.6
repoze.who==2.2
six==1.6.1
speaklater==1.3
sprox==0.9.3
stevedore==0.15
tg.devtools==2.3.2
tgext.admin==0.6.3
tgext.crud==0.7.2
transaction==1.4.3
tw2.core==2.2.1.1
tw2.forms==2.2.0.3
wsgiref==0.1.2
zope.interface==4.1.1
zope.sqlalchemy==0.7.4
waitress
